+++ packages/go-kosu/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## master
 
+-   Add GetBlocks RPC endpoint
 -   Fix wrong RPC remote address in lite mode
 
 ## v0.4.0